Make id field mandatory for entities.
This change has a great implication on how business logic is modeled
with Modelark, as repositories are no longer responsible of assigning
identification to the entities they add. Each entity's ID should now
be generated client side, at the application layer or in a domain service
before been handled to a repository.